refactor: decompose lib/httpapi/server.go into handlers, middleware, routes
Extract handler methods, middleware functions, and route registration
from the monolithic server.go into dedicated files for better
maintainability and code organization.

- handlers.go: all HTTP handler methods (getLogs, getStatus, etc.)
- middleware.go: hostAuthorizationMiddleware, sseMiddleware
- routes.go: registerRoutes, registerStaticFileRoutes, redirectToChat
